MAINT: move dispersion plots and prints
