+++ rocketpy/plots/__init__.py
@@ -1,4 +1,5 @@
 from .compare import *
+from .dispersion_plots import _DispersionPlots
 from .environment_plots import _EnvironmentPlots
 from .flight_plots import _FlightPlots
 from .rocket_plots import _RocketPlots
