+++ rocketpy/plots/dispersion_plots.py
@@ -1,0 +1,177 @@
+__author__ = (
+    "Mateus Stano Junqueira, Guilherme Fernandes Alves, Bruno Abdulklech Sorban"
+)
+__copyright__ = "Copyright 20XX, RocketPy Team"
+__license__ = "MIT"
+
+
+import matplotlib.pyplot as plt
+
+from ..tools import generate_dispersion_ellipses
+
+
+class _DispersionPlots:
+    """Class to plot the dispersion results of the dispersion analysis."""
+
+    def __init__(self, dispersion):
+        self.dispersion = dispersion
+        return None
+
+    def plot_ellipses(
+        self,
+        image=None,
+        actual_landing_point=None,
+        perimeterSize=3000,
+        xlim=(-3000, 3000),
+        ylim=(-3000, 3000),
+        save=False,
+    ):
+        """A function to plot the error ellipses for the apogee and impact
+        points of the rocket. The function also plots the real landing point, if
+        given
+
+        Parameters
+        ----------
+        image : str, optional
+            The path to the image to be used as the background
+        actual_landing_point : tuple, optional
+            A tuple containing the actual landing point of the rocket, if known.
+            Useful when comparing the dispersion results with the actual landing.
+            Must be given in tuple format, such as (x, y) in meters.
+            By default None.
+        perimeterSize : int, optional
+            The size of the perimeter to be plotted. The default is 3000.
+        xlim : tuple, optional
+            The limits of the x axis. The default is (-3000, 3000).
+        ylim : tuple, optional
+            The limits of the y axis. The default is (-3000, 3000).
+        save : bool
+            Whether save the output into a file or not. The default is False.
+            If True, the .show() method won't be called, and the image will be
+            saved with the same name as filename attribute, using a .png format.
+
+        Returns
+        -------
+        None
+        """
+        # Import background map
+        if image is not None:
+            try:
+                from imageio import imread
+
+                img = imread(image)
+            except ImportError:
+                raise ImportError(
+                    "The 'imageio' package could not be. Please install it to add background images."
+                )
+            except FileNotFoundError:
+                raise FileNotFoundError(
+                    "The image file was not found. Please check the path."
+                )
+
+        (
+            impact_ellipses,
+            apogee_ellipses,
+            apogeeX,
+            apogeeY,
+            impactX,
+            impactY,
+        ) = generate_dispersion_ellipses(self.dispersion.results)
+
+        # Create plot figure
+        plt.figure(num=None, figsize=(8, 6), dpi=150, facecolor="w", edgecolor="k")
+        ax = plt.subplot(111)
+
+        for ell in impact_ellipses:
+            ax.add_artist(ell)
+        for ell in apogee_ellipses:
+            ax.add_artist(ell)
+
+        # Draw launch point
+        plt.scatter(0, 0, s=30, marker="*", color="black", label="Launch Point")
+        # Draw apogee points
+        plt.scatter(
+            apogeeX, apogeeY, s=5, marker="^", color="green", label="Simulated Apogee"
+        )
+        # Draw impact points
+        plt.scatter(
+            impactX,
+            impactY,
+            s=5,
+            marker="v",
+            color="blue",
+            label="Simulated Landing Point",
+        )
+        # Draw real landing point
+        if actual_landing_point != None:
+            plt.scatter(
+                actual_landing_point[0],
+                actual_landing_point[1],
+                s=20,
+                marker="X",
+                color="red",
+                label="Measured Landing Point",
+            )
+
+        plt.legend()
+
+        # Add title and labels to plot
+        ax.set_title(
+            "1$\sigma$, 2$\sigma$ and 3$\sigma$ Dispersion Ellipses: Apogee and Lading Points"
+        )
+        ax.set_ylabel("North (m)")
+        ax.set_xlabel("East (m)")
+
+        # Add background image to plot
+        # TODO: In the future, integrate with other libraries to plot the map (e.g. cartopy, ee, etc.)
+        # You can translate the basemap by changing dx and dy (in meters)
+        dx = 0
+        dy = 0
+        if image is not None:
+            plt.imshow(
+                img,
+                zorder=0,
+                extent=[
+                    -perimeterSize - dx,
+                    perimeterSize - dx,
+                    -perimeterSize - dy,
+                    perimeterSize - dy,
+                ],
+            )
+        plt.axhline(0, color="black", linewidth=0.5)
+        plt.axvline(0, color="black", linewidth=0.5)
+        plt.xlim(*xlim)
+        plt.ylim(*ylim)
+
+        # Save plot and show result
+        if save:
+            plt.savefig(
+                str(self.dispersion.filename) + ".png",
+                bbox_inches="tight",
+                pad_inches=0,
+            )
+        else:
+            plt.show()
+        return None
+
+    def plot_results(self):
+        """Plot the results of the dispersion analysis.
+
+        Parameters
+        ----------
+        None
+
+        Returns
+        -------
+        None
+        """
+        for key in self.dispersion.results.keys():
+            plt.figure()
+            plt.hist(
+                self.dispersion.results[key],
+            )
+            plt.title("Histogram of " + key)
+            plt.ylabel("Number of Occurrences")
+            plt.show()
+
+        return None
