+++ rocketpy/prints/__init__.py
@@ -1,4 +1,5 @@
 from .compare_prints import _ComparePrints
+from .dispersion_prints import _DispersionPrints
 from .environment_prints import _EnvironmentPrints
 from .flight_prints import _FlightPrints
 from .parachute_prints import _ParachutePrints
